+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Exposed ispeWidth/ispeHeight to decoder if decoding items with an associated ispe box
 
 ## [0.3.10] - 2019-09-26
 ### Added